phone

+++ fragment.glsl
@@ -1,4 +1,4 @@
-precision mediump float;
+precision highp float;
 uniform float t;
 uniform vec2 resolution;
 uniform sampler2D backBuffer;
@@ -18,7 +18,8 @@ float gold_noise(in vec2 xy, in float seed)
 
 
 void main() {
-  vec2 xy = (uv-vec2(1.0)) * resolution;
+  vec2 xy = (uv-vec2(1.0)) * resolution ;
+  // vec2 xy = fract(uv+vec2(1.0))*2.) * resolution ;
 
   float time = t/ (1000.);
   vec3 color = vec3(gold_noise(xy, fract(time)+1.0), // r
